Extract wrap in version folder
