--- packages/scripts/internationalization/common/error.js
@@ -1,8 +1,0 @@
-const chalk = require('chalk');
-
-function error(message) {
-	console.error(chalk.red(`❌ ${message}`));
-	process.exit(1);
-}
-
-module.exports = error;
