+++ packages/scripts/internationalization/index.js
@@ -2,7 +2,7 @@
 
 const fs = require('fs');
 const path = require('path');
-const error = require('./common/error');
+const { error } = require('./common/log');
 
 module.exports = function runInI18n(command) {
 	function processConfigFile() {
