+++ packages/scripts/internationalization/scripts/to-github.js
@@ -1,24 +1,10 @@
-/* eslint-disable global-require,no-console */
 const fs = require('fs');
 const path = require('path');
 const rimraf = require('rimraf');
-const xmlParser = require('xml2json');
 const spawn = require('cross-spawn');
 
-const error = require('../common/error');
-
-function printSection(title) {
-	console.log('\n------------------------------');
-	console.log(`-- ${title}`);
-	console.log('------------------------------');
-}
-
-function printSuccess(text) {
-	console.log(`✅ ${text}`);
-}
-function printRunning(text) {
-	console.log(`🏃 ${text}`);
-}
+const { error, printSection, printSuccess, printRunning } = require('../common/log');
+const getVersion = require('../common/version');
 
 function getGithubVariables() {
 	const { GITHUB_LOGIN, GITHUB_TOKEN } = process.env;
@@ -52,51 +38,6 @@ function checkI18nFolder(i18nFolder) {
 	} else {
 		printSuccess(`${i18nFolder} found and contains files to add.`);
 	}
-}
-
-function getVersion() {
-	printSection('Extract version');
-	const lernaJsonPath = path.join(process.cwd(), 'lerna.json');
-	const packageJsonPath = path.join(process.cwd(), 'package.json');
-	const pomXmlPath = path.join(process.cwd(), 'pom.xml');
-	const VERSION_REGEX = /^([0-9]+\.[0-9]+).*/;
-
-	let extractedVersion;
-	let source;
-	if (fs.existsSync(lernaJsonPath)) {
-		const { version } = require(lernaJsonPath);
-		const match = version.match(VERSION_REGEX);
-		if (match) {
-			source = 'lerna.json';
-			extractedVersion = match[1];
-		}
-	} else if (!extractedVersion && fs.existsSync(packageJsonPath)) {
-		const { version } = require(packageJsonPath);
-		const match = version.match(VERSION_REGEX);
-		if (match) {
-			source = 'package.json';
-			extractedVersion = match[1];
-		}
-	} else if (!extractedVersion && fs.existsSync(pomXmlPath)) {
-		const data = fs.readFileSync(pomXmlPath);
-		const { version } = xmlParser.toJson(data);
-		const match = version.match(VERSION_REGEX);
-		if (match) {
-			source = 'pom.xml';
-			extractedVersion = match[1];
-		}
-	}
-
-	if (!extractedVersion) {
-		error(`
-			talend-scripts can't determine your project version. Possible reasons : 
-				* no package.json or pom.xml in this directory
-				* no version in package.json or pom.xml
-		`);
-	}
-
-	printSuccess(`Version extracted from ${source}: ${extractedVersion}`);
-	return extractedVersion;
 }
 
 function cloneLocalesRepo({ githubUrl, localesRepoPath }) {
